+++ .rbenv-version
@@ -1,0 +1,1 @@
+ree-1.8.7-2011.03
