+++ COPYING
@@ -1,3 +1,4 @@
+Copyright (c) 2013 Thomas Fuchs
 Copyright (c) 2007-2011 Samuel Tesla
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
