--- test/base32_test.rb
@@ -1,75 +1,0 @@
-# Copyright (c) 2007-2011 Samuel Tesla
-
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-
-# The above copyright notice and this permission notice shall be included in
-# all copies or substantial portions of the Software.
-
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-# THE SOFTWARE.
-
-require 'test/unit'
-require 'base32'
-
-class TestBase32 < Test::Unit::TestCase
-  def assert_decoding(encoded, plain)
-    decoded = Base32.decode(encoded)
-    assert_equal(plain, decoded)
-  end
-
-  def assert_encoding(encoded, plain)
-    actual = Base32.encode(plain)
-    assert_equal(encoded, actual)
-  end
-
-  def assert_encode_and_decode(encoded, plain)
-    assert_encoding(encoded, plain)
-    assert_decoding(encoded, plain)
-  end
-
-  def test_empty_string
-    assert_encode_and_decode('', '')
-  end
-
-  def test_a
-    assert_encode_and_decode('ME======', 'a')
-  end
-
-  def test_12345
-    assert_encode_and_decode('GEZDGNBV', '12345')
-  end
-
-  def test_abcde
-    assert_encode_and_decode('MFRGGZDF', 'abcde')
-  end
-
-  def test_constitution_preamble
-    plaintext =<<-EOT
-      We the people of the United States, in order to form a more perfect union,
-      establish justice, insure domestic tranquility, provide for the common
-      defense, promote the general welfare, and secure the blessings of liberty
-      to ourselves and our posterity, do ordain and establish this Constitution
-      for the United States of America.
-    EOT
-    encoded = %W(
-      EAQCAIBAEBLWKIDUNBSSA4DFN5YGYZJAN5TCA5DIMUQFK3TJORSWIICTORQXIZLTFQQGS3RA
-      N5ZGIZLSEB2G6IDGN5ZG2IDBEBWW64TFEBYGK4TGMVRXIIDVNZUW63RMBIQCAIBAEAQGK43U
-      MFRGY2LTNAQGU5LTORUWGZJMEBUW443VOJSSAZDPNVSXG5DJMMQHI4TBNZYXK2LMNF2HSLBA
-      OBZG65TJMRSSAZTPOIQHI2DFEBRW63LNN5XAUIBAEAQCAIDEMVTGK3TTMUWCA4DSN5WW65DF
-      EB2GQZJAM5SW4ZLSMFWCA53FNRTGC4TFFQQGC3TEEBZWKY3VOJSSA5DIMUQGE3DFONZWS3TH
-      OMQG6ZRANRUWEZLSOR4QUIBAEAQCAIDUN4QG65LSONSWY5TFOMQGC3TEEBXXK4RAOBXXG5DF
-      OJUXI6JMEBSG6IDPOJSGC2LOEBQW4ZBAMVZXIYLCNRUXG2BAORUGS4ZAINXW443UNF2HK5DJ
-      N5XAUIBAEAQCAIDGN5ZCA5DIMUQFK3TJORSWIICTORQXIZLTEBXWMICBNVSXE2LDMEXAU===).join
-    assert_encode_and_decode(encoded, plaintext)
-  end
-end
